fix(tx_pool): compute tx page-pool floor in 64-bit to avoid overflow
homa_tx_pool_gc() turns the tx_page_pool_min_kb floor into a page count
with `homa->tx_page_pool_min_kb * 1000`. tx_page_pool_min_kb is a plain
int, so for large floors (a multi-GB tx reserve is a legitimate sysctl
setting) the multiply overflows 32-bit int. The resulting min_pages is
garbage -- typically negative -- which makes `release = max_low_mark -
min_pages` huge, so gc reclaims pages that the floor was supposed to keep.

Promote the multiply to u64 (and cast the shifted result back to int).

Adds a table-driven regression test,
homa_tx_pool_gc__min_kb_floor_no_overflow, whose corner row sets a floor
whose `* 1000` overflows a 32-bit int: before the fix gc frees the whole
pool (avail 10 -> 0); after the fix the floor is honored and the pool is
left intact. Existing tx_pool suite stays green (30/30).

+++ homa_tx_pool.c
@@ -417,8 +417,8 @@ void homa_tx_pool_gc(struct homa *homa)
 	 * releasing the lock, since freeing is expensive).
 	 */
 	spin_lock_bh(&max_pool->mutex);
-	min_pages = ((homa->tx_page_pool_min_kb * 1000)
-			+ (HOMA_TX_PAGE_SIZE - 1)) >> HOMA_TX_PAGE_SHIFT;
+	min_pages = (int)((((u64)homa->tx_page_pool_min_kb * 1000)
+			+ (HOMA_TX_PAGE_SIZE - 1)) >> HOMA_TX_PAGE_SHIFT);
 
 	/* Note: may need to adjust max_low_mark to reflect changes made
 	 * while lock wasn't held.
